feat: sys模块实体

+++ src/main/java/org/kkx/repository/SysMenuRepository.java
@@ -3,5 +3,5 @@ package org.kkx.repository;
 import org.kkx.entity.SysMenu;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface SysMenuRepository extends PagingAndSortingRepository<SysMenu, Integer> {
+public interface SysMenuRepository extends PagingAndSortingRepository<SysMenu, Long> {
 }
